backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
pydantic-settings==2.2.1
supabase==2.3.6
PyJWT==2.8.0
httpx==0.27.0
